requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.4.4
Flask-Mail==0.9.1
PyMySQL==1.0.3
cryptography==41.0.4

numpy~=2.2.1
PyJWT~=2.10.1
SQLAlchemy~=2.0.36